% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_scale.R
\name{adjust_scale}
\alias{adjust_scale}
\title{Utils: Adjust y and x scale}
\usage{
adjust_scale(
  plots,
  scale.x = "default",
  limits.x = "default",
  scale.y = "default",
  limits.y = "default"
)
}
\arguments{
\item{plots}{Object of analysis or plot_arrange}

\item{scale.x}{x-axis scale (use vector)}

\item{limits.x}{limits in x-axis (use vector)}

\item{scale.y}{y-axis scale (use vector)}

\item{limits.y}{limits in y-axis (use vector)}
}
\value{
Returns the scaled graph
}
\description{
Adjust y and x scale for chart or charts
}
\examples{
library(AgroReg)
data("aristolochia")
attach(aristolochia)
a=LM(trat,resp)
b=LL(trat,resp,npar = "LL.3")
a=plot_arrange(list(a,b),gray = TRUE)
adjust_scale(a,scale.y = seq(0,100,10),limits.y = c(0,100))
}
