% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_scale_x.R
\name{adjust_scale_x}
\alias{adjust_scale_x}
\title{Utils: Adjust x scale}
\usage{
adjust_scale_x(plots, scale = "default", limits = "default")
}
\arguments{
\item{plots}{Object of analysis or plot_arrange}

\item{scale}{x-axis scale (use vector)}

\item{limits}{limits in x-axis (use vector)}
}
\value{
Returns the scaled graph
}
\description{
Adjust x scale for chart or charts
}
\examples{
library(AgroReg)
data("aristolochia")
attach(aristolochia)
a=LM(trat,resp)
b=LL(trat,resp,npar = "LL.3")
a=plot_arrange(list(a,b),gray = TRUE)
adjust_scale_x(a,scale = seq(10,40,5),limits = c(10,40))
}
