% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_scale_y.R
\name{adjust_scale_y}
\alias{adjust_scale_y}
\title{Utils: Adjust y scale}
\usage{
adjust_scale_y(plots, scale = "default", limits = "default")
}
\arguments{
\item{plots}{Object of analysis or plot_arrange}

\item{scale}{y-axis scale (use vector)}

\item{limits}{limits in y-axis (use vector)}
}
\value{
Returns the scaled graph
}
\description{
Adjust y scale for chart or charts
}
\examples{
library(AgroReg)
data("aristolochia")
attach(aristolochia)
a=LM(trat,resp)
b=LL(trat,resp,npar = "LL.3")
a=plot_arrange(list(a,b),gray = TRUE)
adjust_scale_y(a,scale = seq(0,100,10),limits = c(0,100))
}
