% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AgroReg-package}
\alias{AgroReg}
\alias{AgroReg-package}
\title{AgroReg: Regression Analysis Linear and Nonlinear for Agriculture}
\description{
Linear and nonlinear regression analysis common in agricultural science articles (Archontoulis & Miguez (2015). \doi{10.2134/agronj2012.0506}). The package includes polynomial, exponential, gaussian, logistic, logarithmic, segmented, non-parametric models, among others. The functions return the model coefficients and their respective p values, coefficient of determination, root mean square error, AIC, BIC, as well as graphs with the equations automatically.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fisher.uel.br/AgroReg_shiny/}
  \item \url{https://fisher.uel.br/AgroReg_shiny.pt/}
}

}
\author{
\strong{Maintainer}: Gabriel Danilo Shimizu \email{gabrield.shimizu@gmail.com} (\href{https://orcid.org/0000-0001-8524-508X}{ORCID})

Authors:
\itemize{
  \item Leandro Simoes Azeredo Goncalves (\href{https://orcid.org/0000-0001-9700-9375}{ORCID}) [contributor]
}

}
\keyword{internal}
