% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_edition.R
\name{coloredit_arrange}
\alias{coloredit_arrange}
\title{Change the colors of a graph from the plot_arrange function}
\usage{
coloredit_arrange(graphs, color = NA)
}
\arguments{
\item{graphs}{object from a plot_arrange function}

\item{color}{color curve and point}
}
\value{
The function changes the colors of a graph coming from the plot_arrange function
}
\description{
Change the colors of a graph from the plot_arrange function
}
\examples{
library(AgroReg)
data("aristolochia")
attach(aristolochia)
a=LM(trat,resp)
b=LL(trat,resp,npar = "LL.3")
c=BC(trat,resp, npar = "BC.4")
d=CD(trat,resp, npar = "CRS.4")
graph=plot_arrange(list(a,b,c,d))
coloredit_arrange(graph,color=c("black","red","blue","green"))
}
\author{
Gabriel Danilo Shimizu
}
