% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparative_function.R
\name{comparative_model}
\alias{comparative_model}
\title{Analysis: Comparative models}
\usage{
comparative_model(models, names_model = NA, plot = FALSE, round.label = 2)
}
\arguments{
\item{models}{List with objects of type analysis}

\item{names_model}{Names of the models}

\item{plot}{Plot in the parameters}

\item{round.label}{Round label plot}
}
\value{
Returns a table and/or graph with the statistical parameters for choosing the model.
}
\description{
This function allows the construction of a table and/or graph with the statistical parameters to choose the model from the analysis functions.
}
\author{
Gabriel Danilo Shimizu
}
