% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractmodel_analysis.R
\name{extract.model}
\alias{extract.model}
\title{Analysis: Extract models}
\usage{
extract.model(model, type = "model")
}
\arguments{
\item{model}{Object returned from an analysis function}

\item{type}{output type}
}
\value{
Returns an object of class drm, lm or nls (type="model"), or vector of residuals (type="resids"), or graph of the residuals (type="residplot", type="stdresidplot", type=" qqplot").
}
\description{
This function allows extracting the model (type="model") or residuals (type="resids"). The model class depends on the function and can be (lm, drm or nls). This function also allows you to perform graphical analysis of residuals (type="residplot"), graphical analysis of standardized residuals (type="stdresidplot"), graph of theoretical quantiles (type="qqplot").
}
\examples{

data("aristolochia")
attach(aristolochia)
a=linear.linear(trat,resp,point = "mean")
extract.model(a,type = "qqplot")
}
