% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granada_dataset.R
\docType{data}
\name{granada}
\alias{granada}
\title{Dataset: Granada}
\format{
data.frame containing data set
  \describe{
  \item{\code{time}}{numeric vector with times}
  \item{\code{WL}}{Numeric vector with response}
  }
}
\usage{
data("granada")
}
\description{
The data are part of an experiment that studied the drying
kinetics of pomegranate peel over time under an air-circulation
oven. Mass loss was assessed.
}
\examples{
data(granada)
}
\author{
Gabriel Danilo Shimizu
}
\keyword{datasets}
