% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC_analysis.R
\name{interval.confidence}
\alias{interval.confidence}
\title{Analysis: Interval of confidence}
\usage{
interval.confidence(model)
}
\arguments{
\item{model}{Object analysis}
}
\value{
Return in the interval of confidence
}
\description{
Interval of confidence in model regression
}
\examples{
data("granada")
attach(granada)
a=LM(time, WL)
interval.confidence(a)
}
\author{
Gabriel Danilo Shimizu
}
