% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N_analysis.R
\name{Nreg}
\alias{Nreg}
\title{Analysis: Graph for not significant trend}
\usage{
Nreg(
  trat,
  resp,
  ylab = "Dependent",
  xlab = "Independent",
  error = "SE",
  theme = theme_classic(),
  legend.position = "top",
  legend.text = "not~significant",
  legend.add.mean = TRUE,
  legend.add.mean.name = "hat(y)",
  width.bar = NA,
  point = "all",
  textsize = 12,
  add.line = FALSE,
  add.line.mean = FALSE,
  linesize = 0.8,
  linetype = 1,
  pointsize = 4.5,
  pointshape = 21,
  fillshape = "gray",
  colorline = "black",
  fontfamily = "sans",
  print.on = TRUE
)
}
\arguments{
\item{trat}{Numeric vector with dependent variable.}

\item{resp}{Numeric vector with independent variable.}

\item{ylab}{Dependent variable name (Accepts the \emph{expression}() function)}

\item{xlab}{Independent variable name (Accepts the \emph{expression}() function)}

\item{error}{Error bar (It can be SE - \emph{default}, SD or FALSE)}

\item{theme}{ggplot2 theme (\emph{default} is theme_classic())}

\item{legend.position}{legend position (\emph{default} is "top")}

\item{legend.text}{legend text}

\item{legend.add.mean}{Add average in legend}

\item{legend.add.mean.name}{Add media name}

\item{width.bar}{Bar width}

\item{point}{defines whether you want to plot all points ("all") or only the mean ("mean")}

\item{textsize}{Font size}

\item{add.line}{Add line}

\item{add.line.mean}{Add line mean}

\item{linesize}{line size}

\item{linetype}{line type}

\item{pointsize}{shape size}

\item{pointshape}{format point (default is 21)}

\item{fillshape}{Fill shape}

\item{colorline}{Color lines}

\item{fontfamily}{Font family}

\item{print.on}{Print output}
}
\value{
The function returns an exploratory graph of segments
}
\description{
Graph for non-significant trend. Can be used within the multicurve command
}
\examples{
library(AgroReg)
data("aristolochia")
attach(aristolochia)
Nreg(trat,resp)
}
\author{
Gabriel Danilo Shimizu

Leandro Simoes Azeredo Goncalves
}
\keyword{non-significant}
