% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_param.R
\name{stat_param}
\alias{stat_param}
\title{Analysis: Other statistical parameters}
\usage{
stat_param(models, names_model = NA, round = 3)
}
\arguments{
\item{models}{List with objects of type analysis}

\item{names_model}{Names of the models}

\item{round}{Round numbers}
}
\value{
Returns a table with the statistical parameters for choosing the model.
}
\description{
This function calculates other statistical parameters such as Mean (Bias) Error, Relative Mean (Bias) Error, Mean Absolute Error, Relative Mean Absolute Error, Root Mean Square Error, Relative Root Mean Square Error, Modeling Efficiency, Standard deviation of differences, Coefficient of Residual Mass.
}
\examples{
library(AgroReg)
data(granada)
attach(granada)
a=LM(time,WL)
b=LL(time,WL)
c=BC(time,WL)
d=weibull(time,WL)
stat_param(models=list(a,b,c,d))
}
\author{
Gabriel Danilo Shimizu
}
